requirements.txt
grpcio
grpcio-tools
motor
pydantic
python-dotenv
uvloop; sys_platform != "win32"
